Enforce DECIMAL(p,s) precision and scale
DECIMAL(p,s) was parsed and ignored. Now the typmod is captured by the
parser (validated: 1 <= p <= 1000, 0 <= s <= p), stored on the column
descriptor as append-only omitempty JSON fields (zero = bare DECIMAL,
the pre-existing meaning — old descriptors and rolling upgrades are
unaffected; golden descriptor test extended), and enforced with
PostgreSQL semantics at the two row-completion choke points every write
funnels through — the insert-row build (INSERT, COPY, DEFAULT fill) and
the UPDATE SET loop:

- values quantize to scale s with round-half-even (new decimal.Quantize,
  factored from DivQuantize's rounding core);
- integer digits exceeding p-s are SQLSTATE 22003 (new code), checked
  after rounding so 9.999 into DECIMAL(3,2) overflows, PG order;
- quantization runs before key encoding, so DECIMAL primary keys and
  index entries store the rounded value and rounding collisions surface
  as 23505.

Rendering keeps identity and display separate: storage, equality,
grouping, and memo keys still use the canonical trailing-zero-stripped
text, while datums from enforced columns carry the declared scale in a
new display-only Dscale field — Text() pads to it ("9.90"), the binary
NUMERIC encoder derives its wire dscale from the padded render (golden
vectors extended), and RowDescription reports the typmod
((p<<16)|(s+4)) for enforced columns. DEFAULT values are validated at
DDL time. Coerce failures on INSERT/UPDATE that previously surfaced as
internal errors (XX000) now report 22P02, matching COPY.

Tests: decimal.Quantize tie/negative/scale-0 table; parser typmod
capture and rejection table; enforceTypmod unit table (rescale,
post-rounding overflow, boundary, bare/NULL pass-through); two e2e
suites over real pgwire (text + binary formats and binary NUMERIC
params): fixed-scale rendering, ties-to-even, 22003 on INSERT, UPDATE,
COPY, and DDL-time DEFAULT, RowDescription typmod via pgx field
metadata, DECIMAL(6,2) primary key point lookup by any spelling plus
rounding-collision 23505, ORDER BY, secondary index on an enforced
column, COPY rescale. The wire-fidelity e2e suite now uses bare DECIMAL
(typmod behavior has its own suite).

Fixes #39

+++ pkg/testutils/testcluster/decimal_jsonb_test.go
@@ -67,9 +67,10 @@ func TestDecimalJSONBOverPgwire(t *testing.T) {
 	}
 	defer func() { _ = conn.Close(ctx) }()
 
-	// DECIMAL(10,2): precision/scale accepted (and ignored — see docs).
+	// Bare DECIMAL: unconstrained, values survive with full fidelity.
+	// (DECIMAL(p,s) rescaling/enforcement is TestDecimalTypmodOverPgwire.)
 	if _, err := conn.Exec(ctx, `CREATE TABLE items (
-		id INT8 PRIMARY KEY, price DECIMAL(10,2), attrs JSONB
+		id INT8 PRIMARY KEY, price DECIMAL, attrs JSONB
 	)`); err != nil {
 		t.Fatalf("create: %v", err)
 	}
